api-service/apps/tasks/views.py
from rest_framework import viewsets, status
from rest_framework.response import Response
from .models import Task
from .serializers import TaskSerializer
from .services import create_task

class TaskViewSet(viewsets.ModelViewSet):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer

    def create(self, request, *args, **kwargs):
        title = request.data.get('title')
        user_id = request.user.id if request.user.is_authenticated else 1
        
        task = create_task(title=title, user_id=user_id)
        serializer = self.get_serializer(task)
        return Response(serializer.data, status=status.HTTP_201_CREATED)