api-service/requirements.txt
django
djangorestframework
psycopg2-binary
confluent-kafka
pydantic